Long conversion

+++ src/main/java/demo/PCAPMiningStats.java
@@ -2,32 +2,32 @@ package demo;
 
 public class PCAPMiningStats {
 	
-	int pkts_matched;
-	int pkts_searched;
-	int pages_searched;
-	int directory_entries_searched;
-	public int getPkts_matched() {
+	long pkts_matched;
+	long pkts_searched;
+	long pages_searched;
+	long directory_entries_searched;
+	public long getPkts_matched() {
 		return pkts_matched;
 	}
-	public void setPkts_matched(int pkts_matched) {
+	public void setPkts_matched(long pkts_matched) {
 		this.pkts_matched = pkts_matched;
 	}
-	public int getPkts_searched() {
+	public long getPkts_searched() {
 		return pkts_searched;
 	}
-	public void setPkts_searched(int pkts_searched) {
+	public void setPkts_searched(long pkts_searched) {
 		this.pkts_searched = pkts_searched;
 	}
-	public int getPages_searched() {
+	public long getPages_searched() {
 		return pages_searched;
 	}
-	public void setPages_searched(int pages_searched) {
+	public void setPages_searched(long pages_searched) {
 		this.pages_searched = pages_searched;
 	}
-	public int getDirectory_entries_searched() {
+	public long getDirectory_entries_searched() {
 		return directory_entries_searched;
 	}
-	public void setDirectory_entries_searched(int directory_entries_searched) {
+	public void setDirectory_entries_searched(long directory_entries_searched) {
 		this.directory_entries_searched = directory_entries_searched;
 	}
